--- src/main/java/com/example/user/domain/domain/UserRole.java
@@ -1,5 +1,0 @@
-package com.example.user.domain.domain;
-
-public enum UserRole {
-  ANONYMOUS, TESTER, USER, CHIEF, ADMIN
-}
