--- src/main/java/com/example/user/domain/domain/mapper/UserMapper.java
@@ -1,16 +1,0 @@
-package com.example.user.domain.domain.mapper;
-
-import com.example.user.domain.domain.User;
-import com.example.user.domain.domain.UserRole;
-import com.example.user.domain.presentation.dto.request.UserCreateRequest;
-import com.example.user.domain.presentation.dto.response.UserResponse;
-import org.mapstruct.Builder;
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel="spring", builder = @Builder(disableBuilder = false))
-public interface UserMapper {
-
-  User toEntity(UserCreateRequest request, UserRole role);
-
-  UserResponse toDto(User user);
-}
