+++ src/main/java/com/example/user/domain/dto/request/UserNameUpdateRequest.java
@@ -1,0 +1,9 @@
+package com.example.user.domain.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+public record UserNameUpdateRequest(
+        @NotNull(message="name is null")
+        String name
+) {
+}
