+++ src/main/java/com/example/user/domain/dto/request/UserProfileUpdateRequest.java
@@ -1,0 +1,9 @@
+package com.example.user.domain.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+public record  UserProfileUpdateRequest (
+        @NotNull(message="profile is null")
+        String profile
+) {
+}
