+++ src/main/java/com/example/user/domain/dto/response/UserResponse.java
@@ -1,0 +1,9 @@
+package com.example.user.domain.dto.response;
+
+public record UserResponse (
+        String userId,
+        String name,
+        Long remain_token,
+        String profile
+) {
+}
