+++ src/main/java/com/example/user/domain/entity/User.java
@@ -1,0 +1,58 @@
+package com.example.user.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class User {
+  @Id
+  private String userId;
+  @Column(unique = true)
+  private String email;
+  private String name;
+  private String password;
+  @Enumerated(EnumType.STRING)
+  private UserRole role;
+  private Long remain_token;
+  private String profile;
+  @CreationTimestamp
+  private LocalDateTime created_at;
+
+  @PrePersist
+  public void defaultSettings() {
+    if (this.remain_token == null) this.remain_token = 0L;
+    if (this.profile == null) this.profile = "Default.png";
+  }
+
+  public boolean equalsPassword(String password, PasswordEncoder encoder) {
+    return encoder.matches(password, this.password);
+  }
+
+  public void changeToEncodedPassword(String password, PasswordEncoder encoder) {
+    String encodedPassword = encoder.encode(password);
+    changePassword(encodedPassword);
+  }
+
+  private void changePassword(String password) {
+    this.password = password;
+  }
+
+  public void updateProfile(String profile) {
+    this.profile = profile;
+  }
+
+  public void updateName(String name) {
+    this.name = name;
+  }
+}
