+++ src/main/java/com/example/user/domain/entity/UserRole.java
@@ -1,0 +1,5 @@
+package com.example.user.domain.entity;
+
+public enum UserRole {
+  ANONYMOUS, TESTER, USER, CHIEF, ADMIN
+}
