+++ src/main/java/com/example/user/domain/exception/gRPC/GrpcMappedException.java
@@ -1,0 +1,16 @@
+package com.example.user.domain.exception.gRPC;
+
+import org.springframework.http.HttpStatus;
+
+public class GrpcMappedException extends RuntimeException {
+  private final HttpStatus status;
+
+  public GrpcMappedException(String message, HttpStatus status) {
+    super(message);
+    this.status = status;
+  }
+
+  public HttpStatus getStatus() {
+    return status;
+  }
+}
