--- src/main/java/com/example/user/domain/presentation/dto/request/UserNameUpdateRequest.java
@@ -1,9 +1,0 @@
-package com.example.user.domain.presentation.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-
-public record UserNameUpdateRequest(
-        @NotNull(message="name is null")
-        String name
-) {
-}
