--- src/main/java/com/example/user/domain/presentation/dto/request/UserProfileUpdateRequest.java
@@ -1,9 +1,0 @@
-package com.example.user.domain.presentation.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-
-public record  UserProfileUpdateRequest (
-        @NotNull(message="profile is null")
-        String profile
-) {
-}
