--- src/main/java/com/example/user/domain/presentation/dto/response/ResponseDto.java
@@ -1,8 +1,0 @@
-package com.example.user.domain.presentation.dto.response;
-
-public record ResponseDto<T> (
-        String message,
-        T data
-) {
-
-}
