--- src/main/java/com/example/user/domain/presentation/dto/response/UserResponse.java
@@ -1,9 +1,0 @@
-package com.example.user.domain.presentation.dto.response;
-
-public record UserResponse (
-        String userId,
-        String name,
-        Long remain_token,
-        String profile
-) {
-}
