--- src/main/java/com/example/user/domain/service/exception/AlreadyExistsUserEmailException.java
@@ -1,18 +1,0 @@
-package com.example.user.domain.service.exception;
-
-import com.example.user.global.error.Exception.ErrorCode;
-import com.example.user.global.error.Exception.GlobalException;
-
-public class AlreadyExistsUserEmailException extends GlobalException {
-  public AlreadyExistsUserEmailException() {
-    super(ErrorCode.USER_EMAIL_ALREADY_EXISTS);
-  }
-
-  private static class Holder {
-    private static final AlreadyExistsUserEmailException INSTANCE = new AlreadyExistsUserEmailException();
-  }
-
-  public static AlreadyExistsUserEmailException getInstance() {
-    return Holder.INSTANCE;
-  }
-}
