--- src/main/java/com/example/user/domain/service/exception/ValidationPasswordException.java
@@ -1,17 +1,0 @@
-package com.example.user.domain.service.exception;
-
-import com.example.user.global.error.Exception.ErrorCode;
-import com.example.user.global.error.Exception.GlobalException;
-
-public class ValidationPasswordException extends GlobalException {
-  public ValidationPasswordException() {
-    super(ErrorCode.PASSWORD_VALIDATION_FAILED);
-  }
-  private static class Holder {
-    private static final ValidationPasswordException INSTANCE = new ValidationPasswordException();
-  }
-  public static ValidationPasswordException getInstance() {
-    return Holder.INSTANCE;
-  }
-
-}
