--- src/main/java/com/example/user/domain/service/exception/gRPC/GrpcMappedException.java
@@ -1,16 +1,0 @@
-package com.example.user.domain.service.exception.gRPC;
-
-import org.springframework.http.HttpStatus;
-
-public class GrpcMappedException extends RuntimeException {
-  private final HttpStatus status;
-
-  public GrpcMappedException(String message, HttpStatus status) {
-    super(message);
-    this.status = status;
-  }
-
-  public HttpStatus getStatus() {
-    return status;
-  }
-}
