+++ src/main/java/com/example/user/domain/service/gRPC/GrpcUserLoginService.java
@@ -3,8 +3,8 @@ package com.example.user.domain.service.gRPC;
 import com.example.grpc.UserLoginRequest;
 import com.example.grpc.UserLoginResponse;
 import com.example.grpc.UserLoginServiceGrpc;
-import com.example.user.domain.domain.User;
-import com.example.user.domain.domain.repository.UserRepository;
+import com.example.user.domain.entity.User;
+import com.example.user.domain.repository.UserRepository;
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
 import io.grpc.stub.StreamObserver;
