--- src/main/java/com/example/user/global/error/Exception/GlobalException.java
@@ -1,10 +1,0 @@
-package com.example.user.global.error.Exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class GlobalException extends RuntimeException {
-  private final ErrorCode errorCode;
-}
