+++ src/main/java/com/example/user/global/error/exception/GlobalException.java
@@ -1,0 +1,10 @@
+package com.example.user.global.error.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class GlobalException extends RuntimeException {
+  private final ErrorCode errorCode;
+}
