+++ src/main/java/com/example/user/global/mapper/ResponseDtoMapper.java
@@ -1,0 +1,13 @@
+package com.example.user.global.mapper;
+
+import com.example.user.global.mapper.dto.ResponseDto;
+import org.springframework.stereotype.Component;
+
+@Component
+public class ResponseDtoMapper {
+
+  public <T> ResponseDto<T> toResponseDto(String message, T data) {
+    return new ResponseDto<>(message, data);
+  }
+
+}
