+++ src/main/java/com/example/user/global/mapper/dto/ResponseDto.java
@@ -1,0 +1,7 @@
+package com.example.user.global.mapper.dto;
+
+public record ResponseDto<T> (
+        String message,
+        T data
+) {
+}
