+++ src/test/java/com/example/user/domain/service/UserServiceUnitTest.java
@@ -1,14 +1,14 @@
 package com.example.user.domain.service;
 
-import com.example.user.domain.domain.User;
-import com.example.user.domain.domain.UserRole;
-import com.example.user.domain.domain.mapper.UserMapper;
-import com.example.user.domain.domain.repository.UserRepository;
-import com.example.user.domain.service.exception.AlreadyExistsUserEmailException;
-import com.example.user.domain.service.exception.AlreadyExistsUserIdException;
-import com.example.user.domain.service.exception.NotFoundUserException;
-import com.example.user.domain.presentation.dto.request.UserCreateRequest;
-import com.example.user.domain.presentation.dto.request.UserUpdateRequest;
+import com.example.user.domain.entity.User;
+import com.example.user.domain.entity.UserRole;
+import com.example.user.domain.entity.mapper.UserMapper;
+import com.example.user.domain.entity.repository.UserRepository;
+import com.example.user.domain.exception.AlreadyExistsUserEmailException;
+import com.example.user.domain.exception.AlreadyExistsUserIdException;
+import com.example.user.domain.exception.NotFoundUserException;
+import com.example.user.domain.dto.request.UserCreateRequest;
+import com.example.user.domain.controller.dto.request.UserUpdateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
